slowly getting the hang of it.
rename test_helper to helper, so we can use rake to trigger tests

i like some of the stuff I'm seeing in boom, but also some of the stuff I already have,
so taking time to mix it together. getting a handle on rake
I think part of the problem is, things are simpler than i think sometimes.